수정: 테스트 환경 PostgreSQL 전환 + stock_info BigInteger 수정
- 테스트 DB: 인메모리 SQLite → PostgreSQL 16 (프로덕션 동일 DBMS)
- docker-compose.test.yml 신규 (tmpfs, 포트 5433)
- conftest.py: session scope 엔진 + TRUNCATE 격리
- CI: PostgreSQL service container + TEST_DATABASE_URL
- stock_info shares/revenue/operating_income/net_income: Integer → BigInteger

+++ db/models/stock_info.py
@@ -1,6 +1,6 @@
 """StockInfo model — persistent stock info cache (survives Docker restart)."""
 
-from sqlalchemy import Column, Float, Index, Integer, String
+from sqlalchemy import BigInteger, Column, Float, Index, Integer, String
 
 from db.base import Base
 
@@ -14,7 +14,7 @@ class StockInfo(Base):
     change_val = Column(Float)
     change_pct = Column(Float)
     mktcap = Column(Float)
-    shares = Column(Integer)
+    shares = Column(BigInteger)
     high_52 = Column(Float)
     low_52 = Column(Float)
     price_updated_at = Column(String)
@@ -26,9 +26,9 @@ class StockInfo(Base):
     market_type = Column(String)
     sector = Column(String)
     metrics_updated_at = Column(String)
-    revenue = Column(Integer)
-    operating_income = Column(Integer)
-    net_income = Column(Integer)
+    revenue = Column(BigInteger)
+    operating_income = Column(BigInteger)
+    net_income = Column(BigInteger)
     bsns_year = Column(Integer)
     fin_updated_at = Column(String)
     return_3m = Column(Float)
